Release/1.4.2

+++ src/main/java/com/moneyme/moneymebackend/dto/response/TransactionSettlementResponseDTO.java
@@ -15,5 +15,5 @@ public class TransactionSettlementResponseDTO {
     UserResponseDTO from;
     @JsonProperty("to")
     UserResponseDTO to;
-    @JsonProperty("amount") int amount;
+    @JsonProperty("amount") long amount;
 }
